Almost finished

+++ printchars.c
@@ -6,21 +6,21 @@
 /*   By: enunez-n <enunez-n@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/31 11:29:24 by enunez-n          #+#    #+#             */
-/*   Updated: 2022/05/31 12:29:14 by enunez-n         ###   ########.fr       */
+/*   Updated: 2022/05/31 15:50:34 by enunez-n         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_printchar(int c)
+int	ft_printchar(int c)
 {
 	ft_putchar_fd(c, 1);
 	return (1);
 }
 
-size_t	ft_printstring(char *str)
+int	ft_printstring(char *str)
 {
-	size_t	count;
+	int	count;
 
 	if (!str)
 		str = "(null)";
